# backend/docs/openapi/v3-api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Megna Deal Portal API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterRequestDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "lastName": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "companyName": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 8
          }
        },
        "required": [
          "companyName",
          "email",
          "firstName",
          "lastName",
          "password",
          "phone"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "investorId": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          },
          "expiresInSeconds": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}